pyproject.toml
[tool.poetry]
name = "fineberg-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fineberg_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
